+++ mt_metadata/timeseries/filters/coefficient_filter_basemodel.py
@@ -5,11 +5,11 @@ from typing import Annotated
 
 from pydantic import Field
 
-from mt_metadata.base import MetadataBase
+from mt_metadata.timeseries.filters.filter_base_basemodel import FilterBase
 
 
 # =====================================================
-class CoefficientFilter(MetadataBase):
+class CoefficientFilter(FilterBase):
     gain: Annotated[
         float,
         Field(
@@ -17,6 +17,7 @@ class CoefficientFilter(MetadataBase):
             description="Scale factor for a simple coefficient filter.",
             examples="100",
             alias=None,
+            gt=0.0,
             json_schema_extra={
                 "units": None,
                 "required": True,